+++ som_facturacio_switching/__init__.py
@@ -1,10 +1,8 @@
 import facturacio_extra
 import giscedata_polissa
 import giscedata_facturacio
-import giscedata_facturacio_switching
 import giscedata_facturacio_importacio_linia
 import res_municipi
 import wizard
-import giscedata_facturacio_switching
 import giscedata_facturacio_switching_error
